refactor : 수정사항 반영 (#166)
* 중간 저장

* refactor : 수정사항 반영

* fix : 로직 오류 수정

* refactor: N+1 및 Race Condition 방지

* fix : 피드 조회 수정

* feat : 키워드 api 구현

+++ src/main/java/com/example/cp_main_be/domain/social/avatarpost/dto/PostInfoResponse.java
@@ -14,7 +14,7 @@ public record PostInfoResponse(
     String content,
     String imageUrl,
     boolean isLiked,
-    int likeCount,
+    long likeCount,
     int commentCount,
     List<CommentResponseDTO> comments,
     LocalDateTime createdAt,
